Moved summary insights section to its own container
